Improve Input Method and Change Program Language
The previously used gets() methods are replaced with the safer and more powerful
fgets() as recommended. In the process, a few issues arose, which are commented
with TODO tags. Also, the program language is changed to Korean for better
accessibility to Korean users. Possibly, a language option could be added later.