Add Basic GUI Including Main Menu
This commit adds a basic GUI that shows the player a main menu where the player can
access character stats and player name. For now, character stats and player name are
local variables to the main() function, but they can be changed to global variables
to allow functions to access them more easily.